Jeff's Route File
We are now able to pull route/stop points from the database and save
them into a file. We will want to move this to the main activity so
both the stop and route views will simply be able to read from the file
that is populated. Need to change both route and stop view to read from
this file instead of the XML file that was previously being used.