Preliminary list selection and marker changes
The list selection crashes right now, but is close to being done so
that a route or stop can be selected. Marker changes have been started,
just need to finalize the icons (maybe make them higher res) and make
sure we are replacing all of the icons in each view.